requirements.txt
pillow
pytest
pillow_heif